OpenCL: break source in small code units
